Be nice to mobile devices.
Added a viewport and some padding around the body so that the text does not bleed into the edge of the screen.

Note: We could also add a special CSS for small devices to actually put
the menu on top of the .document and make the .document wider. But
that’s no immediate problem at the moment.