automatically create node_list with autodoc.
the node list is automatically created now.
IMPORTANT1: only nodes in mdp.nodes.__all__ are documented!
the order is the same as the one in the __all__ variable.
TODO: re-order nodes so that more important nodes are at the
beginning.
IMPORTANT2: in order for the nodes that are only available when a
dependency is installed to be documented, you need the corresponding
dependency installed!