the recent fix for extapi extension still failed to work.
this fix is uglier than before, and probably it breaks on
every machine which is not mine. but, on the other hand, no one can
actually run epydoc on any other machine. so, all in all, this hack is
not as bad.
when do we leave epydoc and start using a new tool?