Added extension to load mdp descriptions from mdp.
Introduced three directives:

_ long-description-string::
_ middle-description-string::
_ short-description-string::

These load the corresponding text from the mdp module.
the same strings are thus available for the online documentation, for
the setup.py script, i.e. for PyPI, and for the website.

Downside: the link to the api and to the license are lost. Putting them
in the module docs seems awkward to read, maybe once could put the links
at the end of the docstring using the syntax:
`something`_

_ something: http://xxx ?